begin moving sync code into separate modules
